refactor: Update alertdialogtrigger import and export
